mappings
City, Class, Course, Education

+++ spring/src/main/java/ecproject/narvarosystem/Repository/CourseRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface CourseRepository extends CrudRepository<Course, Long> {
+public interface CourseRepository extends CrudRepository<Course, Integer> {
     List<Course> findAll();
     /*
     @Query(value = "SELECT * FROM course co INNER JOIN class cl ON co.EdID = cl.EdID INNER JOIN student st ON st.classID = cl.classID INNER JOIN user us ON us.userID = st.userID WHERE us.userID = :id", nativeQuery= true)
