+++ spring/src/main/java/ecproject/narvarosystem/Repository/EducationRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface EducationRepository extends CrudRepository<Education, Long> {
+public interface EducationRepository extends CrudRepository<Education, Integer> {
     Education findByName(String name);
     List<Education> findAllByName(String name);
     List<Education> findAll();
